+++ x/perp/keeper/perp_test.go
@@ -4,28 +4,24 @@ import (
 	"testing"
 	"time"
 
-	"github.com/NibiruChain/nibiru/collections"
-	"github.com/NibiruChain/nibiru/collections/keys"
-
-	simapp2 "github.com/NibiruChain/nibiru/simapp"
-
 	"github.com/cosmos/cosmos-sdk/simapp"
-
 	sdk "github.com/cosmos/cosmos-sdk/types"
-
-	"github.com/NibiruChain/nibiru/x/common"
-
 	"github.com/stretchr/testify/require"
 
+	"github.com/NibiruChain/nibiru/collections"
+	"github.com/NibiruChain/nibiru/collections/keys"
+	nibisimapp "github.com/NibiruChain/nibiru/simapp"
+	"github.com/NibiruChain/nibiru/x/common"
 	"github.com/NibiruChain/nibiru/x/perp/types"
-	"github.com/NibiruChain/nibiru/x/testutil/sample"
+	"github.com/NibiruChain/nibiru/x/testutil"
 )
 
 func TestKeeperClosePosition(t *testing.T) {
 	// TODO(mercilex): simulate funding payments
 	t.Run("success", func(t *testing.T) {
 		t.Log("Setup Nibiru app, pair, and trader")
-		nibiruApp, ctx := simapp2.NewTestNibiruAppAndContext(true)
+		nibiruApp, ctx := nibisimapp.NewTestNibiruAppAndContext(true)
+		ctx = ctx.WithBlockTime(time.Now())
 		pair := common.MustNewAssetPair("xxx:yyy")
 
 		t.Log("Set vpool defined by pair on VpoolKeeper")
@@ -53,9 +49,9 @@ func TestKeeperClosePosition(t *testing.T) {
 		)
 
 		t.Log("open position for alice - long")
-		ctx = ctx.WithBlockHeight(ctx.BlockHeight() + 1).WithBlockTime(time.Now().Add(time.Minute))
+		ctx = ctx.WithBlockHeight(ctx.BlockHeight() + 1).WithBlockTime(ctx.BlockTime().Add(time.Minute))
 
-		alice := sample.AccAddress()
+		alice := testutil.AccAddress()
 		err := simapp.FundAccount(nibiruApp.BankKeeper, ctx, alice,
 			sdk.NewCoins(sdk.NewInt64Coin("yyy", 300)))
 		require.NoError(t, err)
@@ -75,7 +71,7 @@ func TestKeeperClosePosition(t *testing.T) {
 			CumulativePremiumFractions: []sdk.Dec{
 				sdk.MustNewDecFromStr("0.3")},
 		})
-		bob := sample.AccAddress()
+		bob := testutil.AccAddress()
 		err = simapp.FundAccount(nibiruApp.BankKeeper, ctx, bob,
 			sdk.NewCoins(sdk.NewInt64Coin("yyy", 62)))
 		require.NoError(t, err)