+++ x/perp/module_simulation.go
@@ -29,6 +29,6 @@ func (am AppModule) RandomizedParams(_ *rand.Rand) []simtypes.ParamChange {
 func (am AppModule) RegisterStoreDecoder(_ sdk.StoreDecoderRegistry) {}
 
 // WeightedOperations returns the all the gov module operations with their respective weights.
-func (am AppModule) WeightedOperations(simState module.SimulationState) []simtypes.WeightedOperation {
-	return simulation.WeightedOperations(simState.AppParams, simState.Cdc, am.ak, am.bk, am.keeper)
+func (am AppModule) WeightedOperations(module.SimulationState) []simtypes.WeightedOperation {
+	return simulation.WeightedOperations(am.ak, am.bk, am.keeper)
 }