+++ x/vpool/abci.go
@@ -4,6 +4,9 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	abci "github.com/tendermint/tendermint/abci/types"
 
+	"github.com/NibiruChain/nibiru/collections/keys"
+	"github.com/NibiruChain/nibiru/x/common"
+
 	"github.com/NibiruChain/nibiru/x/vpool/types"
 
 	"github.com/NibiruChain/nibiru/x/vpool/keeper"
@@ -11,15 +14,14 @@ import (
 
 // EndBlocker Called every block to store a snapshot of the vpool.
 func EndBlocker(ctx sdk.Context, k keeper.Keeper) []abci.ValidatorUpdate {
-	for _, pool := range k.GetAllPools(ctx) {
+	for _, pool := range k.Pools.Iterate(ctx, keys.NewRange[common.AssetPair]()).Values() {
 		snapshot := types.NewReserveSnapshot(
 			pool.Pair,
 			pool.BaseAssetReserve,
 			pool.QuoteAssetReserve,
 			ctx.BlockTime(),
-			ctx.BlockHeight(),
 		)
-		k.SaveSnapshot(ctx, snapshot)
+		k.ReserveSnapshots.Insert(ctx, keys.Join(pool.Pair, keys.Uint64(uint64(ctx.BlockTime().UnixMilli()))), snapshot)
 	}
 	return []abci.ValidatorUpdate{}
 }